Code is incomplete
I just figured out that this code while functional does not work as intended, a logical error. It never asks what customer you want to use or if you are a returning customer. It creates customers but never utilizes the customers. HW possibly suggested to not go that far but unlikely.